feat(hono): Instrument middleware handlers in sub-apps (route groups)
