+++ dev-packages/e2e-tests/test-applications/hono-4/tests/middleware.test.ts
@@ -28,7 +28,7 @@ const SCENARIOS = [
     prefix: '/test-subapp-middleware',
     origin: isNode ? OTEL_ORIGIN : MIDDLEWARE_ORIGIN,
   },
-];
+] as const;
 
 for (const { name, prefix, origin } of SCENARIOS) {
   test.describe(name, () => {
@@ -83,7 +83,12 @@ for (const { name, prefix, origin } of SCENARIOS) {
       );
     });
 
-    test.only('multiple middleware are sibling spans under the same parent', async ({ baseURL }) => {
+    test('multiple middleware are sibling spans under the same parent', async ({ baseURL }) => {
+      test.skip(
+        isNode,
+        'Node double-instruments middleware (too many spans) - TODO: fix this in the SDK and re-enable the test',
+      );
+
       const transactionPromise = waitForTransaction(APP_NAME, event => {
         return event.contexts?.trace?.op === 'http.server' && event.transaction === `GET ${prefix}/multi`;
       });
@@ -94,9 +99,8 @@ for (const { name, prefix, origin } of SCENARIOS) {
       const transaction = await transactionPromise;
       const spans = transaction.spans || [];
 
-  // Sort spans because they are in a different order in Node/Bun (OTel-based)
-  const middlewareSpans = spans
-    .sort((a, b) => (a.start_timestamp ?? 0) - (b.start_timestamp ?? 0));
+      // Sort spans because they are in a different order in Node/Bun (OTel-based)
+      const middlewareSpans = spans.sort((a, b) => (a.start_timestamp ?? 0) - (b.start_timestamp ?? 0));
 
       expect(middlewareSpans).toHaveLength(2);
       expect(middlewareSpans[0]?.description).toBe('middlewareA');
