+++ packages/hono/src/bun/middleware.ts
@@ -1,8 +1,8 @@
 import { type BaseTransportOptions, debug, type Options } from '@sentry/core';
 import { init } from './sdk';
 import type { Hono, MiddlewareHandler } from 'hono';
-import { patchAppUse } from '../shared/patchAppUse';
 import { requestHandler, responseHandler } from '../shared/middlewareHandlers';
+import { applyPatches } from '../shared/applyPatches';
 
 export interface HonoBunOptions extends Options<BaseTransportOptions> {}
 
@@ -16,7 +16,7 @@ export const sentry = (app: Hono, options: HonoBunOptions): MiddlewareHandler =>
 
   init(options);
 
-  patchAppUse(app);
+  applyPatches(app);
 
   return async (context, next) => {
     requestHandler(context);
